# save attribute lookup overhead