In the alignmer parser, preload methods `parse`, `append`, `flush` of the mapper class so as to save [object attribute lookup overhead](https://gregoryszorc.com/blog/2019/01/10/what-i%27ve-learned-about-optimizing-python/).

Benchmarks on 100,000 lines of SHOGUN - Bowtie2 alignment:

```
%%timeit -n 100
with open('test.sam', 'r') as f:
    for rmap in parse_align_file(f, Plain()):
        pass
```

before:
```
333 ms ± 279 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

after:
```
328 ms ± 701 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

Therefore, this change saves 5 / 333 ~= **1.5%** runtime.