@qiyunzhu 2020-06-22
Also tested:

1.
```
parse = lambda x: mapper.parse(x, parser=parser)
```

2.
```
from functools import partial
parse = partial(mapper.parse, parser=parser)
```

But they are even slower than the original solution.