@qiyunzhu 2020-06-22
Also refer to Python official optimization guide: [Avoiding dots](https://wiki.python.org/moin/PythonSpeed/PerformanceTips#Avoiding_dots...).